test: move JSONL export tests to base repository test setup
