+++ tests/test_postgreqsl_repository.py
@@ -1,5 +1,6 @@
 import os
 
+from datastew.process.jsonl_adapter import SQLJsonlConverter
 from datastew.repository.postgresql import PostgreSQLRepository
 from tests.base_repository_test_setup import BaseRepositoryTestSetup
 
@@ -13,3 +14,4 @@ class TestPostgreSQLRepository(BaseRepositoryTestSetup):
         super().setUpClass()
         cls.repo_args = (cls.POSTGRES_TEST_URL, cls.vectorizer1)
         cls.repository = PostgreSQLRepository(*cls.repo_args)
+        cls.jsonl_converter = SQLJsonlConverter(dest_dir="test_export")
