+++ tests/test_sqllite_repository.py
@@ -1,3 +1,4 @@
+from datastew.process.jsonl_adapter import SQLJsonlConverter
 from datastew.repository.sqllite import SQLLiteRepository
 from tests.base_repository_test_setup import BaseRepositoryTestSetup
 
@@ -10,3 +11,4 @@ class TestSQLLiteRepository(BaseRepositoryTestSetup):
         super().setUpClass()
         cls.repo_args = ("disk", "sqlite_db", cls.vectorizer1)
         cls.repository = SQLLiteRepository(*cls.repo_args)
+        cls.jsonl_converter = SQLJsonlConverter(dest_dir="test_export")
