+++ tests/test_weaviate_repository.py
@@ -1,3 +1,4 @@
+from datastew.process.jsonl_adapter import WeaviateJsonlConverter
 from datastew.repository.weaviate import WeaviateRepository
 from tests.base_repository_test_setup import BaseRepositoryTestSetup
 
@@ -9,3 +10,4 @@ class TestWeaviateRepository(BaseRepositoryTestSetup):
     def setUpClass(cls):
         super().setUpClass()
         cls.repository = WeaviateRepository(vectorizer=cls.vectorizer1)
+        cls.jsonl_converter = WeaviateJsonlConverter(dest_dir="test_export")
